+++ src/routes/package-show.js
@@ -52,10 +52,32 @@ class PackageShowRoute extends Component {
 
   constructor(props) {
     super(props);
+
+    const {
+      filterTitles,
+      sort,
+      tags,
+      type,
+      'access-type': accessType,
+      selected,
+      searchfield,
+    } = queryString.parse(props.location.search.substring(1));
+
     this.state = {
+      pkgSearchParams: {
+        q: filterTitles,
+        sort,
+        searchfield,
+        filter: {
+          tags,
+          type,
+          selected,
+          'access-type': accessType,
+        }
+      },
       queryId: 0,
-      pkgSearchParams: {},
     };
+
     const { packageId } = props.match.params;
     const [providerId] = packageId.split('-');
     props.getPackage(packageId);
@@ -180,6 +202,24 @@ class PackageShowRoute extends Component {
   }
 
   searchTitles = (pkgSearchParams) => {
+    const { location, history } = this.props;
+    const qs = queryString.parse(location.search, { ignoreQueryPrefix: true });
+    const search = queryString.stringify({
+      ...qs,
+      filterTitles: pkgSearchParams.q,
+      sort: pkgSearchParams.sort,
+      tags: pkgSearchParams.filter?.tags,
+      type: pkgSearchParams.filter?.type,
+      'access-type': pkgSearchParams.filter?.['access-type'],
+      selected: pkgSearchParams.filter?.selected,
+      searchfield: pkgSearchParams.searchfield,
+    });
+
+    history.replace({
+      ...location,
+      search
+    }, { eholdings: true });
+
     this.setState(({ queryId }) => ({
       pkgSearchParams,
       queryId: (queryId + 1)