+++ src/routes/provider-show.js
@@ -42,8 +42,29 @@ class ProviderShowRoute extends Component {
 
   constructor(props) {
     super(props);
+
+    const {
+      filterPackages,
+      sort,
+      tags,
+      type,
+      'access-type': accessType,
+      selected,
+      searchfield,
+    } = queryString.parse(props.location.search);
+
     this.state = {
-      pkgSearchParams: {},
+      pkgSearchParams: {
+        q: filterPackages,
+        sort,
+        searchfield,
+        filter: {
+          tags,
+          type,
+          selected,
+          'access-type': accessType,
+        }
+      },
       queryId: 0,
     };
     const { providerId } = props.match.params;
@@ -79,6 +100,24 @@ class ProviderShowRoute extends Component {
   }
 
   searchPackages = (pkgSearchParams) => {
+    const { location, history } = this.props;
+    const qs = queryString.parse(location.search, { ignoreQueryPrefix: true });
+    const search = queryString.stringify({
+      ...qs,
+      filterPackages: pkgSearchParams.q,
+      sort: pkgSearchParams.sort,
+      tags: pkgSearchParams.filter?.tags,
+      type: pkgSearchParams.filter?.type,
+      'access-type': pkgSearchParams.filter?.['access-type'],
+      selected: pkgSearchParams.filter?.selected,
+      searchfield: pkgSearchParams.searchfield,
+    });
+
+    history.replace({
+      ...location,
+      search
+    }, { eholdings: true });
+
     this.setState(({ queryId }) => ({
       pkgSearchParams,
       queryId: (queryId + 1)